Clean up Platform to be more maintainable
Break up functionality into groups to simplify their configuration and make it easier to see how things change on each platform.

For the interface default method handling, replace reflection on newer JVMs with a multi-release jar and version-based variants of the JVM support. In order to sufficiently test this, the Java tests are now in a sibling module similar to Android, Kotlin, and RoboVM so that they consume the final jar rather than exploded class directories.
