+++ retrofit/build.gradle
@@ -2,6 +2,35 @@ apply plugin: 'java-library'
 apply plugin: 'org.jetbrains.kotlin.jvm'
 apply plugin: 'com.vanniktech.maven.publish'
 
+def addMultiReleaseSourceSet(int version) {
+  def sourceSet = sourceSets.create("java$version")
+  sourceSet.java.srcDir("src/main/java$version")
+
+  // Propagate dependencies to be visible to this version's source set.
+  configurations.getByName("java${version}Implementation").extendsFrom(configurations.getByName('implementation'))
+  configurations.getByName("java${version}Api").extendsFrom(configurations.getByName('api'))
+  configurations.getByName("java${version}CompileOnly").extendsFrom(configurations.getByName('compileOnly'))
+
+  // Allow types in the main source set to be visible to this version's source set.
+  dependencies.add("java${version}Implementation", sourceSets.getByName("main").output)
+
+  tasks.named("compileJava${version}Java", JavaCompile) {
+    javaCompiler = javaToolchains.compilerFor {
+      languageVersion = JavaLanguageVersion.of(version)
+      vendor = JvmVendorSpec.AZUL
+    }
+  }
+
+  tasks.named('jar', Jar) {
+    from(sourceSet.output) {
+      into("META-INF/versions/$version")
+    }
+  }
+}
+
+addMultiReleaseSourceSet(14)
+addMultiReleaseSourceSet(16)
+
 dependencies {
   api libs.okhttp
 
@@ -11,37 +40,11 @@ dependencies {
 
   compileOnly libs.animalSnifferAnnotations
   compileOnly libs.findBugsAnnotations
-
-  testImplementation projects.retrofit.testHelpers
-  testImplementation libs.junit
-  testImplementation libs.truth
-  testImplementation libs.guava
-  testImplementation libs.mockwebserver
 }
 
 jar {
   manifest {
-    attributes  'Automatic-Module-Name': 'retrofit2'
-  }
-}
-
-// Create a test task for each supported JDK.
-(8..21).each { majorVersion ->
-  def jdkTest = tasks.register("testJdk$majorVersion", Test) {
-    javaLauncher = javaToolchains.launcherFor {
-      languageVersion = JavaLanguageVersion.of(majorVersion)
-      vendor = JvmVendorSpec.AZUL
-    }
-
-    description = "Runs the test suite on JDK $majorVersion"
-    group = LifecycleBasePlugin.VERIFICATION_GROUP
-
-    // Copy inputs from normal Test task.
-    def testTask = tasks.getByName("test")
-    classpath = testTask.classpath
-    testClassesDirs = testTask.testClassesDirs
-  }
-  tasks.named("check").configure {
-    dependsOn(jdkTest)
+    attributes 'Automatic-Module-Name': 'retrofit2'
+    attributes 'Multi-Release': 'true'
   }
 }
