+++ retrofit/kotlin-test/build.gradle
@@ -6,6 +6,5 @@ dependencies {
   testImplementation libs.junit
   testImplementation libs.truth
   testImplementation libs.mockwebserver
-  testImplementation libs.kotlin.stdLib
   testImplementation libs.kotlinCoroutines
 }
