+++ retrofit/src/main/java/retrofit2/BuiltInFactories.java
@@ -1,0 +1,52 @@
+/*
+ * Copyright (C) 2024 Square, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package retrofit2;
+
+import static java.util.Arrays.asList;
+import static java.util.Collections.emptyList;
+import static java.util.Collections.singletonList;
+
+import android.annotation.TargetApi;
+import java.util.List;
+import java.util.concurrent.Executor;
+import javax.annotation.Nullable;
+
+class BuiltInFactories {
+  List<? extends CallAdapter.Factory> createDefaultCallAdapterFactories(
+      @Nullable Executor callbackExecutor) {
+    return singletonList(new DefaultCallAdapterFactory(callbackExecutor));
+  }
+
+  List<? extends Converter.Factory> createDefaultConverterFactories() {
+    return emptyList();
+  }
+
+  @TargetApi(24)
+  static final class Java8 extends BuiltInFactories {
+    @Override
+    List<? extends CallAdapter.Factory> createDefaultCallAdapterFactories(
+        @Nullable Executor callbackExecutor) {
+      return asList(
+          new CompletableFutureCallAdapterFactory(),
+          new DefaultCallAdapterFactory(callbackExecutor));
+    }
+
+    @Override
+    List<? extends Converter.Factory> createDefaultConverterFactories() {
+      return singletonList(new OptionalConverterFactory());
+    }
+  }
+}
